website/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Limiter==1.3.1
itsdangerous==1.1.0
Jinja2==2.11.2
limits==1.5.1
MarkupSafe==1.1.1
python-dotenv==0.13.0
six==1.15.0
Werkzeug==1.0.1
gunicorn
flask-bootstrap
requests